+++ b2sdk/transfer/outbound/upload_manager.py
@@ -9,7 +9,6 @@
 ######################################################################
 
 import logging
-import concurrent.futures as futures
 
 from typing import Optional
 
@@ -23,54 +22,24 @@ from b2sdk.file_lock import FileRetentionSetting, LegalHold
 from b2sdk.stream.progress import ReadingStreamWithProgress
 from b2sdk.stream.hashing import StreamWithHash
 from b2sdk.http_constants import HEX_DIGITS_AT_END
-from b2sdk.utils import B2TraceMetaAbstract
 
 from .progress_reporter import PartProgressReporter
+from ..transfer_manager import TransferManager
+from ...utils.thread_pool import ThreadPoolMixin
 
 logger = logging.getLogger(__name__)
 
 
-class UploadManager(metaclass=B2TraceMetaAbstract):
+class UploadManager(TransferManager, ThreadPoolMixin):
     """
     Handle complex actions around uploads to free raw_api from that responsibility.
     """
 
     MAX_UPLOAD_ATTEMPTS = 5
 
-    def __init__(self, services, max_upload_workers=10):
-        """
-        :param b2sdk.v2.Services services:
-        :param int max_upload_workers: a number of upload threads
-        """
-        self.services = services
-
-        self.upload_executor = None
-        self.max_workers = max_upload_workers
-
     @property
     def account_info(self):
         return self.services.session.account_info
-
-    def set_thread_pool_size(self, max_workers):
-        """
-        Set the size of the thread pool to use for uploads and downloads.
-
-        Must be called before any work starts, or the thread pool will get
-        the default size.
-
-        :param int max_workers: maximum allowed number of workers in a pool
-        """
-        if self.upload_executor is not None:
-            raise Exception('thread pool already created')
-        self.max_workers = max_workers
-
-    def get_thread_pool(self):
-        """
-        Return the thread pool executor to use for uploads and downloads.
-        """
-        if self.upload_executor is None:
-            self.upload_executor = futures.ThreadPoolExecutor(max_workers=self.max_workers)
-        return self.upload_executor
 
     def upload_file(
         self,
@@ -84,7 +53,7 @@ class UploadManager(metaclass=B2TraceMetaAbstract):
         file_retention: Optional[FileRetentionSetting] = None,
         legal_hold: Optional[LegalHold] = None,
     ):
-        f = self.get_thread_pool().submit(
+        f = self._thread_pool.submit(
             self._upload_small_file,
             bucket_id,
             upload_source,
@@ -108,7 +77,7 @@ class UploadManager(metaclass=B2TraceMetaAbstract):
         finished_parts=None,
         encryption: EncryptionSetting = None,
     ):
-        f = self.get_thread_pool().submit(
+        f = self._thread_pool.submit(
             self._upload_part,
             bucket_id,
             file_id,