+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+* Introduce a common thread worker pool for all downloads
+
 ### Added
 * Add pypy-3.8 to test matrix
 * Add support for unverified checksum upload mode